% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pubchem_synonyms}
\alias{pubchem_synonyms}
\title{Get synonyms for a compound}
\usage{
pubchem_synonyms(name)
}
\arguments{
\item{name}{Compound name (e.g., "aspirin")}
}
\value{
tibble: cid, synonym
}
\description{
Get synonyms for a compound
}
